+++ Interfaces/GCHP/Chem_GridCompMod.F90
@@ -490,6 +490,22 @@ CONTAINS
 #endif
 
 !
+! !EXPORT STATE:
+!
+    ! Export delta pressure from internal state for first timestep
+    ! pressure scaling before advection, to conserve mass in
+    ! restart file. Use name DELPDRY to not conflict with name
+    ! DELP_DRY used in restart file and internal state.
+    call MAPL_AddExportSpec(GC, &
+       SHORT_NAME         = 'DELPDRY',                       &
+       LONG_NAME          = 'Delta dry pressure across box', &
+       UNITS              = 'hPa',                &
+       PRECISION          = ESMF_KIND_R8,         &
+       DIMS               = MAPL_DimsHorzVert,    &
+       VLOCATION          = MAPL_VLocationCenter, &
+       RC=STATUS  )
+    _VERIFY(STATUS)
+!
 ! !INTERNAL STATE:
 !
 #if defined( MODEL_GEOS )
@@ -1230,6 +1246,8 @@ CONTAINS
     INTEGER                      :: N, trcID
     TYPE(MAPL_MetaComp), POINTER :: STATE => NULL()
     REAL(ESMF_KIND_R8), POINTER  :: Ptr3D(:,:,:) => NULL()
+    REAL(ESMF_KIND_R8), POINTER  :: Ptr3D_int(:,:,:) => NULL()
+    REAL(ESMF_KIND_R8), POINTER  :: Ptr3D_exp(:,:,:) => NULL()
 #endif
 
     ! Internal run alarms
@@ -1685,6 +1703,18 @@ CONTAINS
     ENDIF
 #endif
 
+#ifdef MODEL_GCHPCTM
+    ! Set delta pressure export to interal state variable DELP_DRY.
+    ! This is used in the first timestep in FV3, before advection, to
+    ! adjust species v/v for conservation of restart file mass, if
+    ! delta pressure present in the restart file.
+    call MAPL_GetPointer ( Export, Ptr3d_exp, 'DELPDRY', ALLOC=.TRUE., __RC__ )
+    CALL ESMF_StateGet( INTSTATE, 'DELP_DRY', GcFld, RC=STATUS )
+    CALL ESMF_FieldGet( GcFld, 0, Ptr3D, __RC__ )
+       Ptr3d_exp = Ptr3d * 100.d0
+       Ptr3d => NULL()
+#endif
+
     !=======================================================================
     ! Error trap: make sure that chemistry / emission time step are same and
     ! correspond to the chemistry step set in GEOSCHEMchem_GridComp.rc.
@@ -1902,7 +1932,6 @@ CONTAINS
     ! Initialize carbon coupling / CO production from CO2 photolysis (if used) 
     CALL GEOS_CarbonInit( GC, GeosCF, State_Chm, State_Grid, __RC__ ) 
 
-    !=======================================================================
     ! All done
     !=======================================================================
 #endif
@@ -2695,8 +2724,11 @@ CONTAINS
        ENDIF
 
        !=======================================================================
-       ! On first call, initialize certain State_Chm and State_Met arrays from
-       ! imports if they are found (ewl, 12/13/18)
+       ! On first call, initialize certain State_Chm arrays from
+       ! internal state (restart file) if they are found. Do not initialize
+       ! from met-fields in restart file that were added for post-processing.
+       ! Do not add delta pressure since used for mass conservation scaling
+       ! in FV3 prior advection which comes before GEOS-Chem.
        !=======================================================================
        IF ( FIRST ) THEN
           CALL MAPL_GetPointer( INTSTATE, Ptr3d_R8, 'H2O2AfterChem', notFoundOK=.TRUE., __RC__ )
@@ -2762,32 +2794,12 @@ CONTAINS
           ENDIF
           Ptr2D_R8 => NULL()
 
-          CALL MAPL_GetPointer( INTSTATE, Ptr3d_R8, 'DELP_DRY', notFoundOK=.TRUE., __RC__ )
-          IF ( ASSOCIATED(Ptr3d_R8) .AND. ASSOCIATED(State_Met%DELP_DRY) ) THEN
-             State_Met%DELP_DRY(:,:,1:State_Grid%NZ) =       &
-                                  Ptr3d_R8(:,:,State_Grid%NZ:1:-1)
-          ENDIF
-          Ptr3d_R8 => NULL()
-
-          CALL MAPL_GetPointer( INTSTATE, Ptr3d_R8, 'BXHEIGHT', notFoundOK=.TRUE., __RC__ )
-          IF ( ASSOCIATED(Ptr3d_R8) .AND. ASSOCIATED(State_Met%BXHEIGHT) ) THEN
-             State_Met%BXHEIGHT(:,:,1:State_Grid%NZ) =       &
-                                  Ptr3d_R8(:,:,State_Grid%NZ:1:-1)
-          ENDIF
-          Ptr3d_R8 => NULL()
-
           CALL MAPL_GetPointer( INTSTATE, Ptr3d_R8, 'TSTRAT_ADJ', notFoundOK=.TRUE., __RC__ )
           IF ( ASSOCIATED(Ptr3d_R8) .AND. ASSOCIATED(State_Chm%TStrat_Adj) ) THEN
              State_Chm%TStrat_Adj(:,:,1:State_Grid%NZ) =       &
                                   Ptr3d_R8(:,:,State_Grid%NZ:1:-1)
           ENDIF
-
           Ptr3d_R8 => NULL()
-          CALL MAPL_GetPointer( INTSTATE, Ptr2d_R8, 'TropLev', notFoundOK=.TRUE., __RC__ )
-          IF ( ASSOCIATED(Ptr2d_R8) .AND. ASSOCIATED(State_Met%TropLev) ) THEN
-             State_Met%TropLev = Ptr2d_R8
-          ENDIF
-          Ptr2d_R8 => NULL()
        ENDIF
 #endif
 
@@ -3104,7 +3116,9 @@ CONTAINS
        CALL MAPL_TimerOff(STATE, "CP_AFTR")
 
        ! Update non-species dynamic internal state arrays post-run
-       ! every timestep, except for area which can be set first run only
+       ! every timestep, except for area which can be set first run only.
+       ! Some of the fields are for post-processing but need to be updated
+       ! mid-run for inclusion in mid-run checkpoint files.
        CALL MAPL_GetPointer( INTSTATE, Ptr2d_R8, 'DryDepNitrogen', &
                              notFoundOK=.TRUE., __RC__ )
        IF (ASSOCIATED(Ptr2d_R8) .AND. ASSOCIATED(State_Chm%DryDepNitrogen)) THEN